Rearrange MessageSvc and OnlinePush
